Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
google-cloud-texttospeech==2.16.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.7
python-dotenv==1.0.0
redis==5.0.1
gunicorn==21.2.0
requests==2.31.0
pydantic==2.5.0
markdown-it-py==3.0.0
rich==13.7.0
psutil==5.9.6
pymongo
SQLAlchemy
psycopg2-binary
APScheduler==3.10.0

pymongo
